working_solution/requirements.txt
flask==2.0.1
google-generativeai==0.3.0
pandas==1.5.3
python-docx==0.8.11
reportlab==4.0.4
openpyxl==3.1.2
fpdf2==2.7.5
python-dotenv==1.0.0 